# server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  messages.MessageFriend:
    properties:
      avatar_id:
        type: string
      id:
        type: string
      username:
        type: string
    type: object
  messages.MessageItem:
    properties:
      audio_duration_ms:
        type: integer
      audio_url:
        type: string
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      is_mine:
        type: boolean
      message_type:
        type: string
      read_at:
        type: string
      recipient_id:
        type: string
      sender_id:
        type: string
      track_metadata:
        $ref: '#/definitions/messages.TrackMetadata'
    type: object
  messages.ThreadResponse:
    properties:
      friend:
        $ref: '#/definitions/messages.MessageFriend'
      messages:
        items:
          $ref: '#/definitions/messages.MessageItem'
        type: array
    type: object
  messages.ThreadSummary:
    properties:
      friend:
        $ref: '#/definitions/messages.MessageFriend'
      lastMessage:
        $ref: '#/definitions/messages.MessageItem'
      unreadCount:
        type: integer
    type: object
  messages.TrackMetadata:
    properties:
      activity_id:
        type: string
      album_art_url:
        type: string
      album_name:
        type: string
      artist_name:
        type: string
      spotify_url:
        type: string
      track_id:
        type: string
      track_name:
        type: string
    type: object
...
host: localhost:5000
info:
  contact: {}
  description: API documentation for the Chen, a music compatibility and social platform.
  title: Chen API
  version: "1.0"
paths:
...
  /messages/{friendID}:
    get:
      description: Fetch all messages in a conversation with a specific friend
      parameters:
      - description: Friend's user ID
        in: path
        name: friendID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/messages.ThreadResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get Message Thread
      tags:
      - messages
  /messages/threads:
    get:
      description: Fetch all active message threads for the current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/messages.ThreadSummary'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: List Message Threads
      tags:
      - messages
...
swagger: "2.0"